Support modal transient windows properly in switchWindow
When a window has a modal transient window, that modal will be focused
instead of the real window. This could lead to loops in the window
switching, wher you'd have the modal focused, switch windows to the
original window and get the modal window focused again (essentially
breaking navigation away from the modal).

By resolving a modal transient back to the original window when figuring
out the currently focused window, and skipping them when finding the
next window to focus, we essentially work with non-modal windows only,
which also matches what is shown in the taskbar.